requirements.txt
python-dotenv
dotenv
openai
chromadb
pypdf
langchain
sentence-transformers
pandas
umap-learn
matplotlib